# frontend/openapi.json
{
  "components": {
    "schemas": {
      "Body_import_evidence_bundle": {
        "properties": {
          "bundle": {
            "contentMediaType": "application/octet-stream",
            "description": "Portable Molecule Atlas evidence bundle ZIP",
            "title": "Bundle",
            "type": "string"
          }
        },
        "required": [
          "bundle"
        ],
        "title": "Body_import_evidence_bundle",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "GetRunSummaryOutput": {
        "additionalProperties": false,
        "properties": {
          "capability_id": {
            "const": "get_run_summary",
            "default": "get_run_summary",
            "title": "Capability Id",
            "type": "string"
          },
          "capability_version": {
            "const": "0.1.0",
            "default": "0.1.0",
            "title": "Capability Version",
            "type": "string"
          },
          "contract_version": {
            "const": "0.1.0",
            "default": "0.1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "run": {
            "$ref": "#/components/schemas/RunSummary"
          }
        },
        "required": [
          "correlation_id",
          "run"
        ],
        "title": "GetRunSummaryOutput",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportEvidenceBundleOutput": {
        "additionalProperties": false,
        "properties": {
          "capability_id": {
            "const": "import_evidence_bundle",
            "default": "import_evidence_bundle",
            "title": "Capability Id",
            "type": "string"
          },
          "capability_version": {
            "const": "0.1.0",
            "default": "0.1.0",
            "title": "Capability Version",
            "type": "string"
          },
          "contract_version": {
            "const": "0.1.0",
            "default": "0.1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "idempotency_replayed": {
            "title": "Idempotency Replayed",
            "type": "boolean"
          },
          "run": {
            "$ref": "#/components/schemas/RunSummary"
          }
        },
        "required": [
          "correlation_id",
          "idempotency_replayed",
          "run"
        ],
        "title": "ImportEvidenceBundleOutput",
        "type": "object"
      },
...
      "JsonValue": {},
      "ManifestWarning": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "path": {
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "path"
        ],
        "title": "ManifestWarning",
        "type": "object"
      },
      "MethodSummary": {
        "additionalProperties": false,
        "properties": {
          "adapter_id": {
            "title": "Adapter Id",
            "type": "string"
          },
          "adapter_version": {
            "title": "Adapter Version",
            "type": "string"
          },
          "checkpoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          },
          "checkpoint_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Sha256"
          },
          "command": {
            "items": {
              "type": "string"
            },
            "title": "Command",
            "type": "array"
          },
          "container_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container Digest"
          },
          "container_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container Image"
          },
          "method_id": {
            "title": "Method Id",
            "type": "string"
          },
          "random_seeds": {
            "items": {
              "type": "integer"
            },
            "title": "Random Seeds",
            "type": "array"
          },
          "source_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Commit"
          },
          "upstream_tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Tool"
          },
          "upstream_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Version"
          }
        },
        "required": [
          "method_id",
          "adapter_id",
          "adapter_version",
          "upstream_tool",
          "upstream_version",
          "source_commit",
          "checkpoint_id",
          "checkpoint_sha256",
          "container_image",
          "container_digest",
          "command",
          "random_seeds"
        ],
        "title": "MethodSummary",
        "type": "object"
      },
...
      "RunFailure": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "minLength": 1,
            "title": "Category",
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "stage": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          }
        },
        "required": [
          "category",
          "message",
          "details"
        ],
        "title": "RunFailure",
        "type": "object"
      },
      "RunSummary": {
        "additionalProperties": false,
        "properties": {
          "artifact_count": {
            "minimum": 0.0,
            "title": "Artifact Count",
            "type": "integer"
          },
          "expected_outputs": {
            "items": {
              "type": "string"
            },
            "title": "Expected Outputs",
            "type": "array"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunFailure"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "method": {
            "$ref": "#/components/schemas/MethodSummary"
          },
          "missing_outputs": {
            "items": {
              "type": "string"
            },
            "title": "Missing Outputs",
            "type": "array"
          },
          "prediction_count": {
            "minimum": 0.0,
            "title": "Prediction Count",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "enum": [
              "succeeded",
              "failed",
              "partial",
              "cancelled",
              "unknown"
            ],
            "title": "State",
            "type": "string"
          },
          "validation_counts": {
            "$ref": "#/components/schemas/ValidationCounts"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ManifestWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "schema_version",
          "state",
          "started_at",
          "finished_at",
          "expected_outputs",
          "missing_outputs",
          "failure",
          "method",
          "artifact_count",
          "prediction_count",
          "validation_counts",
          "warnings"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "ValidationCounts": {
        "additionalProperties": false,
        "properties": {
          "error_count": {
            "minimum": 0.0,
            "title": "Error Count",
            "type": "integer"
          },
          "fail_count": {
            "minimum": 0.0,
            "title": "Fail Count",
            "type": "integer"
          },
          "pass_count": {
            "minimum": 0.0,
            "title": "Pass Count",
            "type": "integer"
          },
          "unavailable_count": {
            "minimum": 0.0,
            "title": "Unavailable Count",
            "type": "integer"
          },
          "warning_count": {
            "minimum": 0.0,
            "title": "Warning Count",
            "type": "integer"
          }
        },
        "required": [
          "pass_count",
          "fail_count",
          "warning_count",
          "unavailable_count",
          "error_count"
        ],
        "title": "ValidationCounts",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Candidate review API for small-molecule datasets.",
    "title": "Molecule Atlas API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/evidence/imports": {
      "post": {
        "operationId": "import_evidence_bundle",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_evidence_bundle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportEvidenceBundleOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsafe or invalid evidence bundle"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Import conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Evidence bundle limit exceeded"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Evidence Bundle",
        "tags": [
          "evidence"
        ]
      }
    },
    "/api/evidence/runs/{run_id}": {
      "get": {
        "operationId": "get_run_summary",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRunSummaryOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Evidence run not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Summary",
        "tags": [
          "evidence"
        ]
      }
    },
...
  }
}